📝 댓글 뷰와 모델 추가
