+++ Bamboo/Sources/Comment/CommentView.swift
@@ -1,0 +1,20 @@
+//
+//  CommentView.swift
+//  Bamboo
+//
+//  Created by Mercen on 2023/02/26.
+//
+
+import SwiftUI
+
+struct CommentView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct CommentView_Previews: PreviewProvider {
+    static var previews: some View {
+        CommentView()
+    }
+}
